Weekend Overhaul Update
While my internet was out, I couldn't really do anything else so I decided to fix this application further for a proper demo. This update fixed many of the following:
-src directory cleaned up
-The exam loader now generates properly based on the ID queried in the endpoint
-Quiz is actually submit-able for grade
-Function to grade quiz upon completion
-Upon completing of the quiz, a results page opens based on your performance
-Gauge chart plugin
-api updated to support passing info to frontend.
-logo update, SVG format
-some minor styling udpates
-proper config.js files now instead of having to update each IP in the source code.
-and more

This will probably be the last update while I move focus to other apps. For now, the admin role features are not getting demo'd and will be on the back burner in the meantime.